refactor: Extract common scripts into ProjectDetails
